refactor(report): reorganize the reporter sources
