+++ .eslintignore
@@ -1,7 +1,7 @@
 node_modules
 coverage
 tmp
-src/report/*.template.js
+src/report/**/*.template.js
 test/*/fixtures/**
 test/*/*/fixtures/**
 types/**
