+++ src/report/dot/dot.template.js
