+++ test/report/dot.spec.js
@@ -1,7 +1,7 @@
 "use strict";
 const fs               = require('fs');
 const expect           = require('chai').expect;
-const render           = require('../../src/report/dotReporter');
+const render           = require('../../src/report/dot');
 const deps             = require('./fixtures/cjs-no-dependency-valid.json');
 const unresolvableDeps = require('./fixtures/es6-unresolvable-deps.json');
 const doNotFollowDeps  = require('./fixtures/do-not-follow-deps.json');
