+++ test/report/err.spec.js
@@ -1,6 +1,6 @@
 "use strict";
 const expect   = require('chai').expect;
-const render   = require('../../src/report/errReporter');
+const render   = require('../../src/report/err');
 const okdeps   = require('./fixtures/everything-fine.json');
 const deps     = require('./fixtures/cjs-no-dependency-valid.json');
 const warndeps = require('./fixtures/err-only-warnings.json');
