+++ test/report/html.spec.js
@@ -1,6 +1,6 @@
 "use strict";
 const expect = require('chai').expect;
-const render = require('../../src/report/htmlReporter');
+const render = require('../../src/report/html');
 const deps   = require('./fixtures/cjs-no-dependency-valid.json');
 
 const elFixture = `<!DOCTYPE html>
